// frontend/types/rpc/AttachmentReadyEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AttachmentInfo } from "./AttachmentInfo";

/**
 * `attachment:ready` event payload.
 */
export type AttachmentReadyEvent = { batch_id: string, index: number, info: AttachmentInfo, };